recv: make DATA_READY reflect ssk in-sequence state
In order to make mptcp_poll independent of the subflows, we need
to keep the mptcp DATA_READY flag in sync, i.e., if it is set, at least
one ssk has in-sequence data.

If it is cleared, no further data is available.
Avoid the unconditional clearing on recv entry.
Instead make sure the flag is cleared on exit if there is no more
in-sequence data available.

v2:
 - add back 'done = true' assigment (Paolo)
 - keep 'break' statements instead of 'goto out'. (Paolo)

Signed-off-by: Florian Westphal <fw@strlen.de>