mptcp: avoid data stream corruption on allocation failure.
Currently, when if mpext allocation fails, the data stream
is corrupted, as we can add the required DSS nor roll-back
the TCP status.

Allocate a skb_ext before the actually skb_buff creation/TCP
status update: in case of memory allocation failure we can
bail early, avoiding the above condition.

Additionally, if the ext is unused to to skb collapsing,
cache the ext in the msk for the next send.

Squash-to: "mptcp: Write MPTCP DSS headers to outgoing data packets"
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>