mptcp: Fix dd command line option
"conv=" is a valid command line parameter, but "conf=" is not. The error
message is redirected to /dev/null so it is not obvious when running
tests. The tests seem to pass either way, so it's not clear that even
the "conv=" is really needed.

squashto: mptcp: add basic kselftest for mptcp
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>