mptcp: parse and emit MP_CAPABLE option according to v1 spec.
This implements MP_CAPABLE options parsing and writing according
to the RFC 6824 bis - MPTCP v1.
Local key is sent on syn/ack, and both keys are sent on 3rd ack.
MP_CAPABLE messages len are updated accordingly.
We need the skbuff to emit correctly the above, so we push the skbuff
struct as an argument all the way from tcp code to the relevant mptcp
callbacks.

When processing incoming MP_CAPABLE + data, build a full blown DSS-like
map info, to simplify later processing.
On child socket creation, we need to record the remote key, if available.

Signed-off-by: Christoph Paasch <cpaasch@apple.com>