Squash-to: tcp: Check for filled TCP option space before SACK
Less invasive checks, so that the number of branchs when hitting
the sack code path decreses compared to the current vanilla tree.
--
Some addictional check in tcp_established_options() is needed, otherwise
the tcp header will be corrupted when mptcp_established_options() fully
consumes the TCP option space. This condition can be reached without
MPTCP, so not sure if the result patch is still worthy for -net ?!?

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>