Squash-to: "tcp: Prevent coalesce/collapse when skb has MPTCP extensions"
Allow coalescing when mptcp extensions are equal or 'from' lacks such extensions.

Note: we could be a little less restrictive allowing coalescing with
different data_ack, but this simplify the code a bit, we can improve
later and at least mptcp_net-next do not generate almost identical
DSS with different data_ack.

With the above assumption, we don't need any special action when
moving bits from skb1 to skb2, as the mapping in skb2 will already
fit, and mapping in skb1 will be dropped as/if needed.

When coalescing to a newly allocated skb, we transfer the relevant
extension to. An MPTCP specific helper is added for that goal,
possibly we could use/create a generic one.

Rebased tree available at:
https://github.com/pabeni/mptcp/tree/mptcp_net-next_export_v4.3

v2  -> v3:
 - refine mptcp_ext_matches() checks -> matches if !from_ext
   (was from_ext == to_ext, which was more restrictive)

v1  -> v2:
 - mptcp_skb_ext_move() is a no-op if from lack an MPTCP ext (Christoph)
 - clean-up mptcp_skb_ext_move() as per Florian's suggestions

RFC -> v1:
 - changed helpers checks as per Florian && Mat feedback
 - added comments (Florian)
 - fixed check in tcp_try_coalesce() (Mat)
 - fixed memcpy (Mat && Florian)
 - dropped now unused mptcp_skb_ext_exist()
 - move the mptcp_skb_can_collapse() into the 2nd group of checks in
   tcp_coalese()

This latter point is somewhat relevant. An unpatched kernel allows
collapsing a single skb with 'old' bits (seq lower than 'start'). Than
really means shrink a bit the current skb, possibly without any fourther
action. This change preserve this behavior, but add the restriction to
respect MPTCP coalescing test when trying to collapsing to the next one.
Overall consequences are not 110% clear to me!

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>