ci:upd-tg: let 'set -e' finding errors
In Bash, when the returned status of a function is checked, 'set -e'
doesn't stop the execution sooner, it lets the function continues until
the end or an explicit 'return' statement.

It is then better to remove these:

  <function> || <print error message>

But print the error message in the exit trap (in case of error of
course).

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: I1daf5a53002fe298e0fcb8455f0a9ae85f3b479f