mptcp: make sure timer is inited/started once data is pending for transmission
The mptcp_connect self-tests with additional '-m mmap' argument triggers a
(harmless) WARN_ON_ONCE splat in mptcp_reset_timer() as the mptcp timer
is not initialized yet.

This is because the timer is only set up after the transmit loop, but
consecutive iterations can trigger the splat.

Make sure the timer is set as soon as we pass the first chunk of data.

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/3
Signed-off-by: Florian Westphal <fw@strlen.de>