mptcp: add msk interations helper
mptcp_token_iter_next() allow traversing all the MPTCP
sockets inside the token container belonging to the given
network namespace with a quite standard iterator semantic.

Additionally export mptcp_token_get_sock(), as it also
will be used by the diag module.

That will be used by the next patch, but keep the API generic,
as we plan to use this later for PM's sake.

Reviewed-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Paolo Abeni <pabeni@redhat.com>