bpf_examples: Adding unique mark per subflow example
This patch adds an example of eBPF program able to set different
options per MPTCP subflow, here the socket mark but it is also
possible to set the TCP congestion control algorithm for example.

The subflow detection is still rudimentary as we only may rely on the
is_mptcp field, that is, we have to know the use-case to write a specific
BPF program. Here, we know that the client will create backup subflows
toward the server so we know that a new subflow is identified by the
pair dst_ip:dst_port. This method is unreliable for real use-cases thus
in the future we will have to expose more precise data like the token
from the msk.

Also, in this example, there is no cleanup of the subflow counter, that
will be done in a future patch.

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>