bpf:examples: add generic BPF program loader
In order to avoid to add custom loader for each example of this folder,
a generic loader based on libbpf is added.

It is currently only able to handle BPF_PROG_TYPE_SOCK_OPS programs but it is
easily extensible to other program types.

It will continue to evolve with the future examples added here.

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@tessares.net>