Squash-to: "mptcp: allow picking different xmit subflows"
This improves the current packet scheduler for scenarios
where the msk sockets is using different links with different
bandwith limits.

The commit comment should be changed into:

"""
Update the scheduler to less trivial heuristic: cache
the last used subflow, and try to send on it a reasonably
long burst of data.

When the burst or the subflow send space is exhausted, pick
the subflow with the lower ratio between write space and
send buffer - that is, the subflow with the greater relative
amount of free space.
"""

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>