mptcp: adapt new code to topic
When applying "mptcp: be careful on MPTCP-level ack", it caused some
conflicts in mptcp_send_ack() with "mptcp: protect the rx path with the
msk socket spinlock". The new patch moves some code that was modified by
a previous commit already in the tree.

This commit is there just to make sure the manual modification I did can
be verified easily.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>