Squash-to: "mptcp: re-enable sndbuf autotune"
In the current code there are a bunch of bugs, very
apparent in unidirectional stream tests with a single
flow- which are not covered by the self-test unfortunatelly:

__mptcp_clean_una() can clear SOCK_NOSPACE, since that can happen
before tcp_data_snd_check()/tcp_new_space()/subflow_write_space,
we end-up with no autotune.

subflow_write_space() can clear SOCK_NOSPACE when subflow can send
but not the msk, that is wrong and can cause hangup.

The fix is moving the msk wake in subflow_write_space()

Commit message should be updated replacing:

"""
Let's sendmsg() set the subflows NOSPACE bit when looking for
more memory. Additionally, cleanup the sndbuf propagation from
subflow into the msk, leveraging the subflow write space callback
and dropping a bunch of duplicate code.
"""

with:

"""
Let's sendmsg() set the subflows NOSPACE bit when looking for
more memory and use the subflow write_space callback to propagate
the snd buf update and wake-up the user-space.

Additionally this allows dropping a bunch of duplicate code.
"""

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>