net: fix GSO for SG-enabled devices.
The commit dbd50f238dec ("net: move the hsize check to the else
block in skb_segment") introduced a data corruption for devices
supporting scatter-gatter segmentation.

The problem boils down to signed/unsigned comparison given
unexpected results: if signed 'hsize' is negative, will be considered
greter than a positive 'len', which is unsigned.

This commit addresses the issue explicitly casting len to signed
integer, so that the comparison gives the correct result.

Bisected-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Fixes: dbd50f238dec ("net: move the hsize check to the else block in skb_segment")
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>