mptcp: always graft subflow socket to parent
Currently incoming subflow uses link to the parent socket,
while outgoing one link to a per subflow socket. The latter
is not really needed, except at initial connect() time and
for the first subflow.

Always graft the outgoing subflow to the parent socket and
free the unneeded ones early.

This allows some code cleanup, reduces the amount of memory
used and will simplify the next patch

Reviewed-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Paolo Abeni <pabeni@redhat.com>