squashto: mptcp: schedule worker when subflow is closed
When remote side closes a subflow we should schedule the worker to
dispose of the subflow in a timely manner.

Otherwise, SF_CLOSED event won't be generated until the mptcp
socket itself is closing or local side is closing another subflow.

As noted by Paolo and Matthieu, a subflow that moves to TCP_CLOSE state
might still have data in its rx queue.

Add a helper to only schedule the work queue once the subflow is empty.

For subflows that still have data, we can do same schedule check
in subflow_check_data_avail().

In case we have multiple subflows closing at the same time,
also make the work queue skip subflows that are TCP_CLOSE with
pending data.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>