mptcp: fix DATA_FIN generation on early shutdown.
If the msk is closed before sending or receiving any data,
no DATA_FIN is generated, instead an MPC ack packet is
crafted out.

In the above scenario, the MPTCP protocol create and send a
pure ack and such packets matches also the criteria for an
MPC ack and the protocol tries first to insert MPC options,
leading the described error.

This change address the issue skipping MPC option for DATA_FIN
packets or if the subflow is not established.

To avoid doing multiple times the same test, fetch the data_fin
flag in a bool variable and pass it to both the interested
helpers.

Fixes: 6d0060f600ad ("mptcp: Write MPTCP DSS headers to outgoing data packets")
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Reviewed-by: Mat Martineau <mathew.j.martineau@linux.intel.com>