ci:virtme: get latest tag, not tarball
At the moment, it looks like there is an archive available for v5.11 but
no 'v5.11.0' tag on Git. For the tag, it is normal because Stephen
didn't do any annoucement.

Because we get source code from Git, best to look at the latest
published tags. We could use the REST API from Github but because it is
a mirror, maybe best to look at the main repo. Parsing the HTML is not
really an issue.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>