Squash to "selftests: mptcp: launch mptcp_connect with timeout"
In diag.sh, we want to send signals to mptcp_connect instances that have
been started in the netns. But we cannot send this signal to 'timeout'
otherwise that will stop the timeout and print something like:

  # ./diag.sh: line 1: 9536 Done echo a
  # 9537 User defined signal 1 | timeout ${timeout_test} ip netns (...)

Instead of trying to find the right PID and storing them in an array, we
can simply use the output of 'ip netns pids' and send a signal to all
these PIDs.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>