Squash to "mptcp: clean-up the rtx path"
The MPTCP_RETRANSMIT define is 9 and these tests should be testing if
BIT(9) is set instead of taking the number 9.

This fix is similar to
commit fe202ab685d5 ("mptcp: fix bit MPTCP_PUSH_PENDING tests") by Dan
Carpenter.

Luckily everything was OK before because the only other flag that was
set was "6", using different bits than "9". Note that BIT(6) is also
using different bits than "9".

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>