utils:update: auto-resolve upstreamed commit
If the commit has been upstreamed in a batch of commits modifying the
same area, it creates conflicts with TopGit. It is easy to resolve by
simply taking the version from upstream.

But better to have it done automatically than risking wrong manual
operation by checking if the right commit has been upstreamed.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>