gh: validate: only one commit for stable branch
That will allow us to validate stable branches. We don't need to
validate each commit in this case. Most of the time we will have one or
two commits to validate, it's enough to just check the compilations on
the last commit.

If we want to validate more, we will need to find which commit to
validate, meaning we would have to fetch the stable repo and this takes
time.

Cc: @mjmartineau
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>