ci: cirrus: initial support
The idea is to run all the selftests + kunit + packetdrill tests on a
public CI (cirrus-ci) like it was done internally at Tessares.

From a container having all required tools, a script is launched to
compile the kernel, start a VM using this kernel and the FS of the
container and run the different tests.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>