Squash-to: mptcp: Retransmit DATA_FIN (again)
Previous testing happened to involve cases where data had been sent and
the retrans timer was already running. Packetdrill testing revealed that
the timer was not being started in cases where no data had been
transmitted, so __mptcp_retrans() was never called. This change starts
the timer when the original DATA_FIN is sent, if the timer is not
already running.

Packetdrill PR is forthcoming.

Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Acked-by: Paolo Abeni <pabeni@redhat.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>