gh: irc: notif for TopGit sync failures
Currently, an email is sent but only to the initiator (jenkins-tessares).

Better to have a notif on IRC in case of failure. No need to have one in
case of success because we would already have one when a tag has been
created. And no need to have one if we were already synced with net-next.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>