Squash to "selftests: mptcp: add fullmesh testcases"
To avoid using the >10 numbers for the fullmesh test cases and comments
around to make it a bit more clear this is a workaround, we can pass
'fullmesh_X' where 'X' is the number of addresses we want to add with
the 'fullmesh' flag.

If we detect that the parameter is starting with 'fullmesh_X', we add
the flag and set the corresponding variable to 'X'. Then we continue
like before.

Acked-and-tested-by: Geliang Tang <geliangtang@xiaomi.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>