selftest: wait before trasnmitting first data chunk
Squashto: selftests: mptcp: add mptcp getsockopt test cases

Mathieu reports spurious failure of the selftest:
 mptcp_sockopt.c:362: do_getsockopt_tcp_info: Assertion `ti.ti[0].tcpi_bytes_received == r' failed.
 mptcp_sockopt.c:330: do_getsockopt_mptcp_info: Assertion `s-&gt;mi.mptcpi_rcv_nxt + r == i.mptcpi_rcv_nxt' failed.

This is because client and server are racing.
The first getsockopt() for (MP)TCP_INFO might be called when some data
was already received, but the mptcpi_rcv_nxt assertion relies on the
initial call to fetch a 'start sequence number'.

Extend the pipe (that was used to unblock the 'connect process' when
the server had bound a socket), it now tells when to connet and when
to send the first chunk of data.

In addition, its remote chance, but possible, that tx is slower
than expected so allow tcpinfo to wait a bit for more data to arrive.

After this, 20k invocations pass without failures.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>