gh: irc: send build and sync error msg to #mptcp
These events are exceptional. They should be more visible than in the
middle of all other notifications from the CI because they need a manual
intervention. Also it is good to propagate to more devs there is an
issue with the current export branch.

Moving them to the main channel helps.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>